bugfix #11: Fix metrics not updating when source changed
Refactor how metrics are collected. The metrics were being converted as
soon as they were registered which was causing the value provider function
to be lost. The implementation was changed to only convert the metrics
when they are collected, not when they are registered, to ensure the value
provider is still used to generate new values.