task #9: Refactor how metrics are built
Update the base collector class to refactor how metrics get built. Changing
the code to return a List because it is a more standard pattern and is
much clearer how to implement the method when overriding it. The old way
of calling the `addMetric` method in the base class was too confusing.