bugfix #17: Add test to ensure jetty metrics are updated
Add a new test to ensure the Jetty metrics get updated correctly when
requests are made against a Javalin webserver server. Right now the test is
failing, which confirms this bug.
Re-added the code removed from the server startup process to attempt a fix.
Add new helper method to get web page data as a string.