bugfix #17: Setup the Javalin server to use the correct Jetty server
Found the root cause of this issue. The problem is that one Jetty server was
being created by the metrics server to pull metrics from, but that server
was never hooked into the Javalin server. This meant that requests were
always handled by the Javalin server, which did not trigger the other server
metrics to update.
The fix was to expose the Jetty server used by the metrics server and link
it up to the Javalin instance.