bugfix #17: Add option to set Jetty server in config
Update the config object used to build the metrics server to take the Jetty
server to collect metrics from. If a server is provided, that server will be
used to pull metrics from, otherwise a new server is created.
Update the jetty statistics collector to take a more general ThreadPool
instead of a QueuedThreadPool so that the thread pool can be pulled off an
existing Jetty server.
Refactor how metrics are initialized to make the code easier to read.