task #19: Remove jetty server options from config
Update the metric server config builder, config, and server to remove all
references to the jetty server build options. The code no longer creates
a Jetty server if none was provided. It is not possible to enable Jetty
metrics without providing a Jetty server to pull metrics from.
Remove tests that are now obsolete.